rename hash.h to hashwrapper.h because there is a hash.h in secp256k1.  Although the directory structure shields the build from this conflict, we don't need the confusion of duplicate header file names in a project of this size.
