+++ src/blockrelay/compactblock.cpp
@@ -20,7 +20,7 @@
 #include "consensus/validation.h"
 #include "dosman.h"
 #include "expedited.h"
-#include "hash.h"
+#include "hashwrapper.h"
 #include "main.h"
 #include "net.h"
 #include "parallel.h"
