+++ src/consensus/merkle.cpp
@@ -1,5 +1,5 @@
 #include "merkle.h"
-#include "hash.h"
+#include "hashwrapper.h"
 #include "utilstrencodings.h"
 
 /*     WARNING! If you're reading this because you're learning about crypto
