+++ src/hash.cpp
@@ -3,9 +3,9 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "hash.h"
 #include "crypto/common.h"
 #include "crypto/hmac_sha512.h"
+#include "hashwrapper.h"
 #include "pubkey.h"
