+++ src/hashwrapper.h
