+++ src/primitives/block.cpp
@@ -7,7 +7,7 @@
 #include "primitives/block.h"
 
 #include "crypto/common.h"
-#include "hash.h"
+#include "hashwrapper.h"
 #include "tinyformat.h"
 #include "utilstrencodings.h"
