+++ src/test/hash_tests.cpp
@@ -3,7 +3,7 @@
 // Distributed under the MIT software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "hash.h"
+#include "hashwrapper.h"
 #include "random.h"
 #include "test/test_bitcoin.h"
 #include "utilstrencodings.h"
