+++ src/test/serialize_tests.cpp
@@ -4,7 +4,7 @@
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
 #include "serialize.h"
-#include "hash.h"
+#include "hashwrapper.h"
 #include "streams.h"
 #include "test/test_bitcoin.h"
